Now player can enter the game

+++ game/src/packets/to_client/extended/auto_soulshots.rs
@@ -1,14 +1,15 @@
 use l2_core::shared_packets::write::SendablePacketBuffer;
+use macro_common::SendablePacket;
 
 #[allow(unused)]
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, SendablePacket)]
 pub struct AutoSoulShots {
     pub(crate) buffer: SendablePacketBuffer,
 }
 
 impl AutoSoulShots {
     const PACKET_ID: u8 = 0xFE;
-    const EX_PACKET_ID: u8 = 0x0C;
+    const EX_PACKET_ID: u16 = 0x0C;
 
     pub fn new(item_id:i32, enable:bool, the_type:i32) -> anyhow::Result<Self> {
         let mut inst = Self {
